refactor: 모든 서비스 로직을 LottoService 클래스로 통합
