+++ src/main/java/lotto/Application.java
@@ -1,9 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.model.LottoRankingService;
 import lotto.model.LottoService;
-import lotto.model.PrizeService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -13,9 +11,7 @@ public class Application {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
         LottoService lottoService = new LottoService();
-        LottoRankingService lottoRankingService = new LottoRankingService();
-        PrizeService prizeService = new PrizeService();
-        LottoController lottoController = new LottoController(inputView, outputView, lottoService, lottoRankingService, prizeService);
+        LottoController lottoController = new LottoController(inputView, outputView, lottoService);
 
         lottoController.run();
     }
