+++ src/main/java/lotto/controller/LottoController.java
@@ -14,21 +14,11 @@ public class LottoController {
     private final InputView inputView;
     private final OutputView outputView;
     private final LottoService lottoService;
-    private final LottoRankingService lottoRankingService;
-    private final PrizeService prizeService;
 
-    public LottoController(
-            InputView inputView,
-            OutputView outputView,
-            LottoService lottoService,
-            LottoRankingService lottoRankingService,
-            PrizeService prizeService
-    ) {
+    public LottoController(InputView inputView, OutputView outputView, LottoService lottoService) {
         this.inputView = inputView;
         this.outputView = outputView;
         this.lottoService = lottoService;
-        this.lottoRankingService = lottoRankingService;
-        this.prizeService = prizeService;
     }
 
     public void run() {
@@ -36,8 +26,7 @@ public class LottoController {
         PurchaseAmount purchaseAmount = readPurchaseAmount();
 
         // [기능] 구입 금액에 해당하는 만큼 로또 발행
-        lottoService.purchase(purchaseAmount);
-        List<Lotto> lottos = lottoService.getLottos();
+        List<Lotto> lottos = lottoService.purchase(purchaseAmount);
 
         // [출력] 구매 개수와 구매한 로또 리스트
         List<LottoDto> lottoDtos = mapToLottoDtos(lottos);
@@ -47,12 +36,13 @@ public class LottoController {
         // [입력] 당첨 번호와 보너스 번호
         WinningNumbers winningNumbers = readWinningNumbers();
         BonusNumber bonusNumber = readBonusNumber();
+        lottoService.checkDuplicate(winningNumbers, bonusNumber);
 
         // [기능] 등수 배열 임시 생성
         int[] rawRanks = new int[7];
         lottos.forEach(lotto -> {
             // [기능] 번호 일치 여부에 따라 등수 매기기
-            int rank = lottoRankingService.determineRank(lotto, winningNumbers, bonusNumber);
+            int rank = lottoService.determineRank(lotto, winningNumbers, bonusNumber);
             rawRanks[rank]++;
         });
 
@@ -66,11 +56,11 @@ public class LottoController {
         long totalPrize = 0L;
         for (int i = 1; i <= 5; i++) {
             if (ranks.get(i) != 0) {
-                long prize = prizeService.payPrize(i);
+                long prize = lottoService.payPrize(i);
                 totalPrize += prize;
             }
         }
-        String profitRate = lottoRankingService.calculateProfitRate(totalPrize, totalPurchase);
+        String profitRate = lottoService.calculateProfitRate(totalPrize, totalPurchase);
 
         // [출력] 당첨 통계
         outputView.showWinningStatisticsTitle();
