+++ src/test/java/lotto/model/LottoServiceTest.java
@@ -1,6 +1,8 @@
 package lotto.model;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.List;
 
@@ -40,9 +42,91 @@ class LottoServiceTest {
         PurchaseAmount purchaseAmount = new PurchaseAmount(3000);
 
         // when
-        lottoService.purchase(purchaseAmount);
+        List<Lotto> lottos = lottoService.purchase(purchaseAmount);
 
         // then
-        assertThat(lottoService.getLottos()).hasSize(3);
+        assertThat(lottos).hasSize(3);
+    }
+
+    @Test
+    void 로또_번호가_당첨_번호_6개와_일치하면_1등이다() {
+        // given
+        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
+        BonusNumber bonusNumber = new BonusNumber(7);
+
+        // when
+        int rank = lottoService.determineRank(lotto, winningNumbers, bonusNumber);
+
+        // then
+        assertThat(rank).isEqualTo(1);
+    }
+
+    @Test
+    void 로또_번호가_당청_번호_5개와_보너스_번호와_일치하면_2등이다() {
+        // given
+        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 7));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
+        BonusNumber bonusNumber = new BonusNumber(7);
+
+        // when
+        int rank = lottoService.determineRank(lotto, winningNumbers, bonusNumber);
+
+        // then
+        assertThat(rank).isEqualTo(2);
+    }
+
+    @Test
+    void 로또_번호가_당첨_번호_5개와_일치하면_3등이다() {
+        // given
+        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 11));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
+        BonusNumber bonusNumber = new BonusNumber(7);
+
+        // when
+        int rank = lottoService.determineRank(lotto, winningNumbers, bonusNumber);
+
+        // then
+        assertThat(rank).isEqualTo(3);
+    }
+
+    @Test
+    void 로또_번호가_당첨_번호_4개와_일치하면_4등이다() {
+        // given
+        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 11, 12));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
+        BonusNumber bonusNumber = new BonusNumber(7);
+
+        // when
+        int rank = lottoService.determineRank(lotto, winningNumbers, bonusNumber);
+
+        // then
+        assertThat(rank).isEqualTo(4);
+    }
+
+    @Test
+    void 로또_번호가_당첨_번호_3개와_일치하면_5등이다() {
+        // given
+        Lotto lotto = new Lotto(List.of(1, 2, 3, 11, 12, 13));
+        WinningNumbers winningNumbers = new WinningNumbers(List.of(1, 2, 3, 4, 5, 6));
+        BonusNumber bonusNumber = new BonusNumber(7);
+
+        // when
+        int rank = lottoService.determineRank(lotto, winningNumbers, bonusNumber);
+
+        // then
+        assertThat(rank).isEqualTo(5);
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {"1,2000000000", "2,30000000", "3,1500000", "4,50000", "5,5000"})
+    void 등수에_따라_당첨_금액을_부여한다(int rank, long winningPrize) {
+        // given by parameter
+
+        // when
+        long prize = lottoService.payPrize(rank);
+
+        // then
+        assertThat(prize).isEqualTo(winningPrize);
     }
 }
